refactor: move tests - webstorm issue
