+++ packages/analytics/src/tests/index.spec.ts
@@ -2,9 +2,9 @@
  * @vitest-environment jsdom
  */
 
-import {initOrbiter} from './index';
-import {PerformanceServices} from './services/performance.services';
-import {UserAgentServices} from './services/user-agent.services';
+import {initOrbiter} from '../index';
+import {PerformanceServices} from '../services/performance.services';
+import {UserAgentServices} from '../services/user-agent.services';
 
 vi.mock('./utils/window.env.utils', () => ({
   envSatelliteId: vi.fn(() => 'satellite-from-env'),
