+++ packages/analytics/src/tests/utils/date.utils.spec.ts
@@ -1,4 +1,4 @@
-import {nowInBigIntNanoSeconds} from './date.utils';
+import {nowInBigIntNanoSeconds} from '../../utils/date.utils';
 
 describe('date.utils', () => {
   describe('nowInBigIntNanoSeconds', () => {
