+++ packages/analytics/src/tests/utils/env.utils.spec.ts
@@ -1,4 +1,4 @@
-import {isBrowser} from './env.utils';
+import {isBrowser} from '../../utils/env.utils';
 
 describe('env.utils', () => {
   afterAll(() => {
